Decimal cumulated return fix.

+++ Sources/Analysis.API/Models/Data/Return.cs
@@ -3,6 +3,6 @@
     public class Return
     {
         public int Year { get; set; }
-        public double CumulatedReturn { get; set; }
+        public decimal CumulatedReturn { get; set; }
     }
 }
